+++ next-docs/public/examples/iconButton/Animations.tsx
@@ -1,24 +1,12 @@
 import { IconButton } from '@heathmont/moon-core-tw';
-import { GenericSettings } from '@heathmont/moon-icons-tw';
+import { OtherFrame } from '@heathmont/moon-icons-tw';
 
 const Example = () => (
   <>
-    <IconButton
-      animation="progress"
-      icon={<GenericSettings className="text-moon-24" />}
-    />
-    <IconButton
-      animation="success"
-      icon={<GenericSettings className="text-moon-24" />}
-    />
-    <IconButton
-      animation="error"
-      icon={<GenericSettings className="text-moon-24" />}
-    />
-    <IconButton
-      animation="pulse"
-      icon={<GenericSettings className="text-moon-24" />}
-    />
+    <IconButton animation="progress" icon={<OtherFrame />} />
+    <IconButton animation="success" icon={<OtherFrame />} />
+    <IconButton animation="error" icon={<OtherFrame />} />
+    <IconButton animation="pulse" icon={<OtherFrame />} />
   </>
 );