+++ workspaces/core/src/button/private/utils/buttonAnimations/AnimationContent.tsx
@@ -1,6 +1,11 @@
 import React from 'react';
 import { GenericCheckAlternative } from '@heathmont/moon-icons-tw';
 import Loader from '../../../../loader/Loader';
+import mergeClassnames from '../../../../mergeClassnames/mergeClassnames';
+import IconLeft from '../../../styles/IconLeft';
+import IconRight from '../../../styles/IconRight';
+import getIconSize from '../buttonSizes/getIconSize';
+import getLoaderSize from '../buttonSizes/getLoaderSize';
 import type ButtonSettingsProps from '../../types/ButtonSettingsProps';
 
 const AnimationContent: React.FC<ButtonSettingsProps> = ({
@@ -9,21 +14,34 @@ const AnimationContent: React.FC<ButtonSettingsProps> = ({
   iconRight,
   iconOnly,
   animation,
+  size,
+  fullWidth,
 }) => (
-  <span className="block relative h-full pointer-events-none">
-    <span className="flex absolute top-1/2 left-1/2 translate-y-[-50%] translate-x-[-50%] content-center justify-center">
-      {animation === 'progress' && <Loader color="currentColor" size="xs" />}
+  <span className="block h-full pointer-events-none">
+    <span
+      className={mergeClassnames(
+        'flex absolute top-1/2 left-1/2 translate-y-[-50%] translate-x-[-50%] content-center',
+        'justify-center'
+      )}
+    >
+      {animation === 'progress' && (
+        <Loader color="currentColor" size={getLoaderSize(size)} />
+      )}
       {animation === 'success' && (
         <GenericCheckAlternative
           aria-label="Success"
-          className="text-moon-24"
+          className={getIconSize(size)}
         />
       )}
     </span>
-    <span className="block opacity-0">
-      {iconLeft}
+    <span className="flex gap-2 items-center opacity-0">
+      {iconLeft && (
+        <IconLeft fullWidth={fullWidth} iconLeft={iconLeft} size={size} />
+      )}
       {children}
-      {iconRight}
+      {iconRight && (
+        <IconRight fullWidth={fullWidth} iconRight={iconRight} size={size} />
+      )}
       {iconOnly}
     </span>
   </span>