+++ workspaces/core/src/button/styles/Hover.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import mergeClassnames from '../../utils/mergeClassnames';
+import mergeClassnames from '../../mergeClassnames/mergeClassnames';
 
 type Props = {
   isHover: boolean;
@@ -9,7 +9,8 @@ const Hover: React.FC<Props> = ({ isHover }) => (
   <span
     aria-hidden="true"
     className={mergeClassnames(
-      'z-[-1] block absolute inset-0 pointer-events-none transition-[background-color_0.2s_ease-in-out]',
+      'z-[-1] block absolute inset-0 pointer-events-none',
+      'transition-[background-color_0.2s_ease-in-out]',
       isHover && 'bg-bulma/[.07]'
     )}
   />