+++ workspaces/core/src/popover/private/types/CallableChildren.ts
@@ -1,0 +1,13 @@
+import { ReactNode } from 'react';
+
+type CallableChildren = (data: {
+  open?: boolean;
+  close?: (
+    focusableElement?:
+      | HTMLElement
+      | React.MutableRefObject<HTMLElement | null>
+      | undefined
+  ) => void;
+}) => ReactNode;
+
+export default CallableChildren;