+++ workspaces/core/src/popover/private/utils/PopoverContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+import type PopoverState from '../types/PopoverState';
+
+const PopoverContext = createContext<PopoverState>({});
+PopoverContext.displayName = 'PopoverContext';
+
+export default PopoverContext;