+++ packages/draggabletable/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 9.18.0
+
+### Minor Changes
+
+- Deploy minor version
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-components@9.18.0
+  - @heathmont/moon-table@9.18.0
+  - @heathmont/moon-themes@9.18.0
+  - @heathmont/moon-utils@9.18.0
+
 ## 9.17.0
 
 ### Minor Changes