+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 9.18.0
+
+### Minor Changes
+
+- Deploy minor version
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-themes@9.18.0
+
 ## 9.17.0
 
 ### Minor Changes